% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_milliseconds.R
\name{add_milliseconds}
\alias{add_milliseconds}
\title{Add Milliseconds to a SubRip (SRT) Timestamp}
\usage{
add_milliseconds(x, ms)
}
\arguments{
\item{x}{Vector of timestamps.}

\item{ms}{Numeric vector of milliseconds to add.}
}
\value{
A character vector corresponding to x.
}
\description{
Convenience method for directly adding milliseconds to a timestamp.
Equivalent to \code{as_timestamp(as_milliseconds(x) + ms)}.
}
