% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_milliseconds.R
\name{as_milliseconds}
\alias{as_milliseconds}
\title{Convert a SubRip (SRT) Timestamp to Milliseconds}
\usage{
as_milliseconds(x)
}
\arguments{
\item{x}{Character vector of SubRip timestamps.}
}
\value{
A numeric vector corresponding to x.
}
\description{
Convert a vector of SubRip-style timestamps into milliseconds. Checks
that each element is of the form hours:minutes:seconds,milliseconds
where units are fixed as two zero-padded digits and the fractional
component is fixed as three zero-padded digits (i.e., \code{00:00:00,000}).
}
