% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_timestamp.R
\name{as_timestamp}
\alias{as_timestamp}
\title{Convert Milliseconds to a SubRip (SRT) Timestamp}
\usage{
as_timestamp(x)
}
\arguments{
\item{x}{Vector of timestamps.}
}
\value{
A character vector corresponding to x.
}
\description{
Convert Milliseconds to a SubRip (SRT) Timestamp
}
