% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/downloadMost_Subtitle.R
\name{downloadMost_Subtitle}
\alias{downloadMost_Subtitle}
\title{Function downloads the newest, the most recommended or the most popular subtitles from opensubtitles.org}
\usage{
downloadMost_Subtitle(which = "popular", unzip = TRUE, info = FALSE)
}
\arguments{
\item{which,}{character one of `popular`, `new` or `recommended`}

\item{unzip,}{logical, if TRUE the .zip file is unzipped}

\item{info,}{logical, if TRUE the file with information is unzipped, otherwise it is deleted}
}
\description{
Function creates folder with downloaded subtitles (as .zip file or unzipped .txt or .srt file)
}
\examples{
downloadMost_Subtitle()
downloadMost_Subtitle('new')
downloadMost_Subtitle(unzip=FALSE)
downloadMost_Subtitle('recommended', unzip=TRUE, info=TRUE)
}
\author{
Justyna Jankowiak
}

