% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fak_jankowiak.R
\name{fak_jankowiak}
\alias{fak_jankowiak}
\title{Rysowanie kolorowej chmury slow z napisow}
\usage{
fak_jankowiak(tytul, jezyk)
}
\arguments{
\item{tytul}{zmienna typu character, case-insensitive, wskazujaca napisy jakiego filmu nas interesuja.
Tytuly filmow nie musza byc w oryginalnym jezyku.}
\item{jezyk}{szukanych napisow.}
}
\value{
wdzieczy i kolorowy obrazek
}
\description{
Funkcja \code{fak_jankowiak} pobiera do folderu napisy do filmow pod wskazanym jezykiem.
}
\details{
Pobieranie pierwszych z brzegu napisow z portalu opensubtitles i rysowanie ich w postaci kolorowej chmury.
Funkcja jest poczatkiem czegos wielkiego: mozna rozbudowac ja dowolnie, by uzytkownik wskazal ile/ktore napisy go
konkretnie interesuja.
}
\examples{
fak_jankowiak("La vita e bella","pol")
fak_jankowiak("Madagaskar","jpn") # napisy w jezyku japonskim.
}
\author{
Justyna Jankowiak & Katarzyna Fak
}
