% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/film_info.R
\name{film_info}
\alias{film_info}
\title{Pobieranie informacji o filmie}
\usage{
film_info(link)
}
\arguments{
\item{link}{zmienna typu character wskazujaca na sciezke do filmu w serwisie IMDb.}
}
\value{
folder z zawierajacy dane o filmie
}
\description{
Funkcja \code{film_info} pobiera do folderu:
1. informacje o filmie (tytul, rok, rezyseria, muzyka itd...) (z portalu IMDb)
2. uproszczone statystyki dotyczace filmu: ocena, liczba oddanych glosow (z portalu IMDb)
3. napisy do filmu (ze strony opensubtitles.org)
4. informacje o tym, czy film dostal nagrode
}
\examples{
film_info("http://www.imdb.com/title/tt0118799/")
}
\author{
opracowala: Katarzyna Fak na podstawie projektow grupy FSS i PRS i wlasnej pracy domowej 06.
}

