% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetMovieInfo.R
\name{GetMovieInfo}
\alias{GetMovieInfo}
\title{Pobieranie informacji o filmie}
\usage{
GetMovieInfo(title)
}
\arguments{
\item{title}{tytul filmu w jezyku angielskim. Im dokladniejszy tym lepsze
jakosciowo informacje o filmie otrzymamy}
}
\value{
Wszystkie pobrane informacje sa zapisywane w pliku tekstowym o nazwie takiej
jak argument \code{title} w folderze roboczym.
}
\description{
Funkcja \code{GetMovieInfo} pobiera informacje o filmie z takich serwisow jak:
IMDB, Wikipedia, OpenSubtitles
}
\details{
Funkcja poprzez podany \code{title} wyszukuje film w serwisie IMDB.
Skad pobiera:
      tytul polski,
      tytul angielski,
      gatunki,
      rok produkcji,
      budzet,
      opis,
      ocene,
      kraje produkcji,
      rezyserow,
      scenarzystow,
      aktorow,
      producentow.
Nastepnie stara sie znalezc ten film na angielskiej wikipedii skad pobiera
analogiczne dane.
Na koniec funkcja wyszukuje film w serwisie OpenSubtitles i pobiera link
do sciagniecia napisow.
}
\examples{
GetMovieInfo("Inception")
GetMovieInfo("Lord of the rings")
GetMovieInfo("Harry Potter")
}
\author{
Karolina Wyszynska
}

