% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imdb_wikipedia_subtitles_info.R
\name{imdb_wikipedia_subtitles_info}
\alias{imdb_wikipedia_subtitles_info}
\title{Funkcja pobiera informacje z IMDB, WIKIPEDII i linki do polskich napisow z OPENSUBTITLER.}
\usage{
imdb_wikipedia_subtitles_info(url)
}
\arguments{
\item{url}{- bezposredni link do filmu ze strony imdb.com}
}
\value{
lista parametrow z kazdej strony
}
\description{
Funkcja pobiera informacje z IMDB, WIKIPEDII i linki do polskich napisow z OPENSUBTITLER.
}
\examples{
imdb_wikipedia_subtitles_info(url = "http://www.imdb.com/title/tt0095016/")
}
\author{
Pawel grabowski
}

