% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NewSubtitleList.R
\name{NewSubtitleList}
\alias{NewSubtitleList}
\title{Information about newest subtitles on opensubtitles.org}
\usage{
NewSubtitleList(language = "all")
}
\arguments{
\item{language}{character vector, the language, in which subtitles will
be searched}
}
\description{
Function \code{NewSubtitleList} return data.frame with information about newest subtitles
on opensubtitles.org
}
\examples{
NewSubtitleList()
NewSubtitleList('eng')
NewSubtitleList('pol')
}
\author{
Mikolaj Wasniewski
}

