% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ostatni_wpis_na_blogu.R
\name{ostatni_wpis_na_blogu}
\alias{ostatni_wpis_na_blogu}
\title{Pobierz ostatni wpis na blogu .subtitler.org}
\usage{
ostatni_wpis_na_blogu()
}
\value{
wektor napisow
}
\description{
Funkcja \code{ostatni_wpis_na_blogu} pobiera ostatni wpis na blogu i komentarze do niego.
}
\author{
Marta Sommer
}

