% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/OstatnieRomunskieNapisy.R
\name{OstatnieRomunskieNapisy}
\alias{OstatnieRomunskieNapisy}
\title{Ostatnio dodane napisy w jezyku romunskim.}
\usage{
OstatnieRomunskieNapisy(hd = FALSE)
}
\arguments{
\item{hd}{Zmienna logiczna, jesli TRUE to zwraca napisy tylko dla filmow 'w duzej rozdzielczosci', domyslnie FALSE.}
}
\description{
Funkcja \code{OstatnieRomunskieNapisy()} patrzy na stronie ostatnio dodane napisy do romunskich filmow i zapisuje je do pliku. Kazdy tytul filmu jest zapisywany z jego rokiem premiery.
}
\examples{
OstatnieRumunskieNapisy('example.txt')
OstatnieRumunskieNapisy('example.txt', hd = TRUE)
}
\author{
Pawel Grabowski
}

