% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pobierz_info_just_jankowiak.R
\name{pobierz.info.just.jankowiak}
\alias{pobierz.info.just.jankowiak}
\title{Funkcja pobiera informacje ze strony IMDb oraz Wikipedii o danym filmie, oraz pobiera napisy do niego.}
\usage{
pobierz.info.just.jankowiak(id)
}
\arguments{
\item{id,}{id filmu (zgodnie z IMDb)}
}
\description{
Funkcja pobiera informacje ze strony IMDb oraz Wikipedii o danym filmie, oraz pobiera napisy do niego.
}
\examples{
pobierz.info.just.jankowiak(0499549) #Avatar
pobierz.info.just.jankowiak(0110912) #Pulp Fiction#'
pobierz.info.just.jankowiak(0126029) #Titanic
pobierz.info.just.jankowiak(0050083) #Dwunastu gniewnych ludzi
}
\author{
Justyna Jankowiak
}

