% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pobierz_info_rudask.R
\name{pobierz_info_rudas_k}
\alias{pobierz_info_rudas_k}
\title{Funkcja zapisuje plik z napisami dla danego filmu i znajduje informacje o nim.}
\usage{
pobierz_info_rudas_k(id, lan = "eng")
}
\arguments{
\item{id,}{id filmu (zgodnie z opensubtitles.org)}

\item{lan}{jezyk, w ktorym maja byc napisy (domyslnie angielski)}
}
\description{
Funkcja \code{pobierz_info_rudas_k(id,lan)} pobiera napisy ze strony http://www.opensubtitles.org/ a takze znajduje informacje o filmie z Wikipedii i z IMDb
}
\examples{
pobierz_info_rudas_k(1) #Obcy 3
pobierz_info_rudas_k(123) #Schaolin Soccer
pobierz_info_rudas_k(2553) #Tora! Tora! Tora!
pobierz_info_rudas_k(31497,"pol") #Jak rozpetalem druga wojne swiatowa
}
\author{
Krzysztof Rudas
}

