% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pobierz_napisy.R
\name{pobierz_napisy}
\alias{pobierz_napisy}
\title{Pobieranie napisow}
\usage{
pobierz_napisy(tytul, jezyk)
}
\arguments{
\item{tytul}{zmienna typu character, case-insensitive, wskazujaca napisy jakiego filmu nas interesuja.
Tytuly filmow nie musza byc w oryginalnym jezyku.}

\item{jezyk}{szukanych napisow.}
}
\value{
folder z zawierajacy napisy do filmow z oryginalna nazwa pobrana z opensubtitles.
}
\description{
Funkcja \code{pobierz_napisy} pobiera do folderu napisy do filmow pod wskazanym jezykiem.
}
\details{
Pobieranie napisow z portalu opensubtitles nie jest trywialna sprawa. Link przekierowujacy
uzyskany funkcja \code{znajdz_tytul_jezyk}() nie zawsze przekierowuje od razu pod adres z ktorego mozna
pobrac zipa z napisami: czesto nalezy kliknac jeszcze raz w link z filmem. Funkcja obsluguje ten przypadek.
Ponadto uzytkownik nie jest zmuszony podawac pelna nazwe filmu, jesli jej nie pamieta, np. wystarczy ze poda
w tytule "Shrek" a program sam dopyta czy chodzilo o "Shrek 2", "Shrek Forever" czy starego dobrego
Shreka, ktory dopiero uwalnia ksiezniczke z wiezy.
}
\examples{
pobierz_napisy("La vita e bella","pol")
pobierz_napisy("Madagaskar","jpn")  # napisy w jezyku japonskim.
}
\author{
Katarzyna Fak
}

