% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pobierz_napisy_do_chmury.R
\name{pobierz_napisy_do_chmury}
\alias{pobierz_napisy_do_chmury}
\title{Pobieranie napisow do chmury slow}
\usage{
pobierz_napisy_do_chmury(tytul, jezyk)
}
\arguments{
\item{tytul}{zmienna typu character, case-insensitive, wskazujaca napisy jakiego filmu nas interesuja.
Tytuly filmow nie musza byc w oryginalnym jezyku.}
\item{jezyk}{szukanych napisow.}
}
\value{
plik tekstowy z napisami
}
\description{
Funkcja \code{pobierz_napisy_do_chmury} pobiera do folderu napisy do filmow pod wskazanym jezykiem.
}
\details{
Pobieranie pierwszych z brzegu napisow z portalu opensubtitles.
}
\examples{
pobierz_napisy("La vita e bella","pol")
pobierz_napisy("Madagaskar","jpn") # napisy w jezyku japonskim.
}
\author{
Katarzyna Fak
}
