% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_srt.R
\name{read_srt}
\alias{read_srt}
\title{Read a SubRip (SRT) file into a tibble}
\usage{
read_srt(file)
}
\arguments{
\item{file}{Path to a file.}
}
\value{
A data frame.
}
\description{
Reads a SubRip format file (extension \code{.srt}) and creates a
tibble from it, with rows corresponding to subtitle records in the
file.
}
