% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readSubtitle.R
\name{readSubtitle}
\alias{readSubtitle}
\alias{subtitle}
\title{Download subtitles from a given link}
\usage{
readSubtitle(link, ...)
}
\arguments{
\item{link}{A link [URL].}

\item{...}{Other parameter passed to readLines().}
}
\value{
Character vector with the file's entry.
}
\description{
Function \code{readSubtitle} reads directly from http the subtitle in srt format.
}
\examples{
readSubtitle('http://dl.opensubtitles.org/pl/download/file/1954081967')
\dontrun{
readSubtitle()
}
}
\seealso{
\code{\link{readLines}} for other arguments.
}

