% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/smuda_sudol.R
\name{smuda_sudol}
\alias{smuda_sudol}
\title{Pobieranie i wyznaczenie czestosci liter w napisach}
\usage{
smuda_sudol(film, jezyk)
}
\arguments{
\item{film}{zmienna typu character, case-insensitive, wskazujaca napisy jakiego filmu nas interesuja.
Tytuly filmow nie musza byc w oryginalnym jezyku.}

\item{jezyk}{szukanych napisow.}
}
\value{
Zwraca wektor nazwany z liczba wystapien danych liter z nazw.
}
\description{
Funkcja \code{smuda_sudol} pobiera do folderu napisy do filmu ze wskazanym jezykiem oraz wyznacza czestosc
liter dla tych napisow
}
\details{
Funkcja pobiera napisy do filmu przy pomocy funkcji \code{pobierz_napisy}, a nastepnie wczytuje
napisy z folderu z napisami i zlicza czestosc uzytych liter w tych napisach.
}
\examples{
smuda_sudol("Titanic","pol")
}
\author{
Adrianna Sudol, Piotr Smuda
}

