% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stat_czas.R
\name{stat_czas}
\alias{stat_czas}
\title{Sprawdz czas wyswietlania napisow w filmie.}
\usage{
stat_czas(link, sciezka)
}
\arguments{
\item{link}{Wektor napisow dlugosci 1. Link do pobrania napisow do filmow pochodzacych ze strony opensubtitles.org.}

\item{sciezka}{Wektor napisow dlugosci 1. Sciezka do miejsca, w ktorym zapisane zostana napisy.}
}
\value{
Wektor numeryczny dlugosci 6 zawierajacy podstawowe statystyki dla czasu wyswietlania napisow w danym filmie.
}
\description{
Funkcja \code{stat_czas} zapisuje napisy do filmow oraz zwraca podstawowe statystyki dla dlugosci wyswietlania napisow.
}
\details{
Jezeli argument \code{sciezka} nie zostanie podany to napisy do filmu nie zostana zapisane. Funkcja \code{stat_czas} rysuje dodatkowo czas wyswietlania sie napisow w zaleznosci od czasu trwania filmu z wyrysowanym splinem.
}
\examples{
link <- "http://dl.opensubtitles.org/pl/download/file/1954081967"
sciezka <- stri_paste(getwd(), "/film_przykladowy.txt")
statystyki <- stat_czas(link = link, sciezka = sciezka)
}
\author{
Marcin Rdzanowski
}

