% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtitler.R
\docType{package}
\name{subtitler}
\alias{subtitler}
\alias{subtitler-package}
\title{subtitler}
\description{
Read and write video subtitle formats.
}
\details{
See the README on \href{https://github.com/benjcunningham/subtitler}{GitHub}.
}
