% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Wasniewski_Grabowski.R
\name{Wasniewski_Grabowski}
\alias{Wasniewski_Grabowski}
\title{Chmura słów dla napisów filmu}
\usage{
Wasniewski_Grabowski(tytul, jezyk, liczbaSlow = 100)
}
\arguments{
\item{tytul}{argument typu character dlugosci 1, tytul filmu (lub jego czesc) dla ktorego ma zostac zrobiony wordcloud}

\item{jezyk}{argument typu character dlugosci 1, jezyk dla ktorego ma zostac zrobiony wordcloud}

\item{liczbaSlow}{liczba słóœ jaka ma się znaleść w obrazku}
}
\description{
Funkcja \code{Wasniewski_Grabowski} zwraca obrazek z chmura słów dla napisow z podanego filmu w danym języku.
Obrazek jest zapisywany w working directory w formacie png.
}
\examples{
Wasniewski_Grabowski('goblet of fire', 'eng', 200)
Wasniewski_Grabowski('skyfall', 'eng')
}
\author{
Mikołaj Waśniewski, Paweł Grabowski
}

