% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_srt.R
\name{write_srt}
\alias{write_srt}
\title{Write a SubRip (SRT) File}
\usage{
write_srt(x, path, append = FALSE)
}
\arguments{
\item{x}{A data frame to write to disk. Must contain the columns
\code{index}, \code{start}, \code{end}, and \code{text}.}

\item{path}{Path to write to.}

\item{append}{If FALSE, will overwrite existing file. If TRUE, will
append to existing file. In both cases, if file does not exist a
new file is created.}
}
\value{
The input x, invisibly.
}
\description{
Takes a data frame of SubRip subtitle components and writes them to
file to the specification of the format.
}
