% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wyszynska_rudas.R
\name{wyszynska_rudas}
\alias{wyszynska_rudas}
\title{Funkcja \code{wyszynska_rudas} buduje chume slow z napisow do filmu podanego jako tytul}
\usage{
wyszynska_rudas(tytul, jezyk)
}
\arguments{
\item{tytul}{tytul filmu}

\item{jezyk}{jezyk, w  ktorym maja zostac pobrane napisy (patrz wymagania funkcji pobierz_napisy)}
}
\description{
Funkcja \code{wyszynska_rudas} buduje chume slow z napisow do filmu podanego jako tytul
}
\examples{
wyszynska_rudas("Lord of the Rings", "eng")
wyszynska_rudas("lego przygoda", "pol")
wyszynska_rudas("The Ring", "eng")
wyszynska_rudas("Kraina Lodu", "eng")
wyszynska_rudas("Shrek", "pol")
}
\author{
Karolina Wyszynska & Krzysztof Rudas
}

