% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/znajdz.R
\name{znajdz}
\alias{znajdz}
\title{Znajdowanie napisow do dowolnego filmu poprzez strone www.opensubtitles.org}
\usage{
znajdz(nazwa_filmu)
}
\arguments{
\item{nazwa_filmu}{Wektor napisow z nazwa szukanego flmu.}
}
\value{
sciezka do strony interenetowej gdzie znajduja sie napisy szukanego flimu
}
\description{
Funkcja \code{znajdz} znajduje sciezke do strony gdzie sa napisy do szukanego filmu.
}
\examples{
znajdz('titanic')
znajdz('Ala ma kota')
}
\author{
Adrianna Sudol
}

