% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/znajdz_info_o_filmie.R
\name{znajdz_info_o_filmie}
\alias{znajdz_info_o_filmie}
\title{Pobierz charakterystyki filmow z roznych portali.}
\usage{
znajdz_info_o_filmie(tytul, rok_produkcji, w_jakim_folderze_zapisac,
  jaki_portal)
}
\arguments{
\item{tytul}{tytul filmu}

\item{rok_produkcji}{rok_produkcji filmu}

\item{w_jakim_folderze_zapisac}{sciezka do folderu, w ktorym chcemy zapisac pobrane dane (zapisza sie w formacie uproszczonytytul_imdb.txt, uproszczonytytul_opensubtitles.txt lub uproszczonytytul_wikipedia.txt)}

\item{jaki_portal}{wektor napisow (do wyboru: "imdb", "wikipedia", "opensubtitles")}
}
\value{
Funkcja \code{znajdz_info_o_filmie} zapisuje maksymalnie trzy pliki .txt (kazdy zawiera ramke danych) w wybranym przez nas folderze.
}
\description{
Funkcja \code{znajdz_info_o_filmie} szuka informacji o filmach na Wikipedii, Opensubtitles lub IMDB.
}
\author{
Marta Sommer
}

