% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zrob_chmure.R
\name{zrob_chmure}
\alias{zrob_chmure}
\title{Rysowanie kolorowej chmury slow z napisow}
\usage{
zrob_chmure(tytul, jezyk)
}
\arguments{
\item{tytul}{zmienna typu character, case-insensitive, wskazujaca napisy jakiego filmu nas interesuja.
Tytuly filmow nie musza byc w oryginalnym jezyku.}
\item{jezyk}{szukanych napisow.}
}
\value{
wdzieczy i kolorowy obrazek
}
\description{
Funkcja \code{zrob_chmure} pobiera do folderu napisy do filmow pod wskazanym jezykiem.
}
\details{
Pobieranie pierwszych z brzegu napisow z portalu opensubtitles i rysowanie ich w postaci kolorowej chmury.
Funkcja jest poczatkiem czegos wielkiego: mozna rozbudowac ja dowolnie, by uzytkownik wskazal ile/ktore napisy go
konkretnie interesuja.
}
\examples{
zrob_chmure("La vita e bella","pol")
zrob_chmure("Madagaskar","jpn") # napisy w jezyku japonskim.
}
\author{
Justyna Jankowiak & Katarzyna Fak
}
