% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_milliseconds.R
\name{as_milliseconds}
\alias{as_milliseconds}
\title{Convert a SubRip (SRT) Timestamp to Milliseconds}
\usage{
as_milliseconds(x)
}
\arguments{
\item{x}{Character vector of timestamps in the form
\code{HH:MM:SS,mmm}.}
}
\value{
A numeric vector corresponding to x.
}
\description{
Convert a SubRip (SRT) Timestamp to Milliseconds
}
