% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_srt.R
\name{read_srt}
\alias{read_srt}
\title{Read a SubRip (SRT) File}
\usage{
read_srt(file)
}
\arguments{
\item{file}{Path to a file.}
}
\value{
A data frame.
}
\description{
Read a SubRip format (filename extension \code{.srt}) text file.
}
