% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subtitler.R
\docType{package}
\name{subtitler}
\alias{subtitler}
\alias{subtitler-package}
\title{subtitler: An Easy Access to the OpenSubtitles Database}
\description{
subtitler: An Easy Access to the OpenSubtitles Database
}

